Fixbug ASROUTER-1386

+++ ET2500/vpp-24.02/src/vnet/dpo/lookup_dpo.h
@@ -100,7 +100,7 @@ typedef struct lookup_dpo_t
     /**
      * Number of locks
      */
-    u16 lkd_locks;
+    u32 lkd_locks;
 } lookup_dpo_t;
 
 extern void lookup_dpo_add_or_lock_w_fib_index(fib_node_index_t fib_index,
